+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Gitaly changelog
 
-UNRELEASED
+[v0.85.0](https://gitlab.com/gitlab-org/gitaly/tags/v0.85.0)
 
 - Implement recursive tree entries fetching
   https://gitlab.com/gitlab-org/gitaly/merge_requests/600